Added the boolean option "Enable Overriding Template World Settings", which defaults to false, which operates just like previous versions. However, if set to true, the user's specified world creation settings will override whatever was in the original template.
This was desired by RockoBonaparte for his Baby's First Space Race modpack. The
plan was to start new users in a alternate tutorial dimension (DIM=2) using a
template. However, the Overworld had been deleted. There is a command block that
will send them to their Overworld spawn, and that Overworld is to be generated
dynamically based on the whatever settings the user gave when they created the
world.

Modpacks that want something like a "lobby dimension" on top of a regular game
would also take advantage of this.

The code targets 1.7, using the 1.7.2 codebase, which is compatible with 1.7.10.