This is a naive port of the 1.7 enable terrain generation override patch for Lockdown to 1.8. Most things were a 1-1 translation. However, it is important to note the new feature does not appear to completely work correctly. If a player creates a new world from a save that has this feature enabled, a new overworld is generated while other dimensions are preserved, but the player will start in that overworld, rather in the dimension where the player had saved the template. This is different from the 1.7 patch, and it is unclear why.
This is the commit information for the 1.7 change for a better explanation of what we are trying to do.

Author:		rocko <rockobonaparte@hotmail.com>
Date:		1 day ago (Wed Dec 23 14:31:31 2015)
Commit hash:	fa814b94a2e718ab3b8eb3ae0d3837db5098337a
Parent(s):	2b3ab41a26

Added the boolean option "Enable Overriding Template World Settings", which
defaults to false, which operates just like previous versions. However, if set
to true, the user's specified world creation settings will override whatever was
in the original template.

This was desired by RockoBonaparte for his Baby's First Space Race modpack. The
plan was to start new users in a alternate tutorial dimension (DIM=2) using a
template. However, the Overworld had been deleted. There is a command block that
will send them to their Overworld spawn, and that Overworld is to be generated
dynamically based on the whatever settings the user gave when they created the
world.

Modpacks that want something like a "lobby dimension" on top of a regular game
would also take advantage of this.

The code targets 1.7, using the 1.7.2 codebase, which is compatible with 1.7.10.